Major refactoring to be more singular
