+++ xhydro/optimal_interpolation/constants.py
@@ -1,4 +1,4 @@
 import os
 
 ROOT_PATH = os.path.abspath(os.path.dirname(__file__))
-DATA_PATH = ROOT_PATH + '\\donnees\\'
+DATA_PATH = ROOT_PATH + '\\data\\'
